requirements.txt
pyyaml
luxai-magpie[discovery,audio]
